parametrizing tests tokenization, cleaner tests
