--- tests/test_tokenize_one_track.py
@@ -1,115 +1,0 @@
-#!/usr/bin/python3 python
-
-"""One track test file
-"""
-
-from copy import deepcopy
-from pathlib import Path
-from typing import Any, Dict, Optional, Sequence, Tuple, Union
-
-import pytest
-from symusic import Score
-
-import miditok
-
-from .utils import (
-    ALL_TOKENIZATIONS,
-    MIDI_PATHS_ONE_TRACK,
-    TEST_LOG_DIR,
-    TOKENIZER_CONFIG_KWARGS,
-    adjust_tok_params_for_tests,
-    prepare_midi_for_tests,
-    tokenize_and_check_equals,
-)
-
-default_params = deepcopy(TOKENIZER_CONFIG_KWARGS)
-default_params.update(
-    {
-        "use_chords": False,  # set false to speed up tests
-        "use_rests": True,
-        "use_tempos": True,
-        "use_time_signatures": True,
-        "use_sustain_pedals": True,
-        "use_pitch_bends": True,
-        "use_pitch_intervals": True,
-        "log_tempos": True,
-        "chord_unknown": False,
-        "delete_equal_successive_time_sig_changes": True,
-        "delete_equal_successive_tempo_changes": True,
-        "sustain_pedal_duration": True,
-    }
-)
-TOK_PARAMS_ONE_TRACK = []
-for tokenization_ in ALL_TOKENIZATIONS:
-    params_ = deepcopy(default_params)
-    adjust_tok_params_for_tests(tokenization_, params_)
-    TOK_PARAMS_ONE_TRACK.append((tokenization_, params_))
-
-
-@pytest.mark.parametrize("midi_path", MIDI_PATHS_ONE_TRACK)
-def test_one_track_midi_to_tokens_to_midi(
-    midi_path: Union[str, Path],
-    tok_params_sets: Optional[Sequence[Tuple[str, Dict[str, Any]]]] = None,
-    saving_erroneous_midis: bool = True,
-):
-    r"""Reads a MIDI file, converts it into tokens, convert it back to a MIDI object.
-    The decoded MIDI should be identical to the original one after downsampling, and
-    potentially notes deduplication. We only parametrize for midi files, as it would
-    otherwise require to load them multiple times each.
-    # TODO test parametrize params_set --> unique method for test tok (one+multi)
-
-    :param midi_path: path to the MIDI file to test.
-    :param tok_params_sets: sequence of tokenizer and its parameters to run.
-    :param saving_erroneous_midis: will save MIDIs decoded with errors, to be used to
-        debug.
-    """
-    if tok_params_sets is None:
-        tok_params_sets = TOK_PARAMS_ONE_TRACK
-    at_least_one_error = False
-
-    # Reads the midi
-    midi = Score(midi_path)
-    # Will store the tracks tokenized / detokenized, to be saved in case of errors
-    for ti, track in enumerate(midi.tracks):
-        track.name = f"original {ti} not quantized"
-    tracks_with_errors = []
-
-    for tok_i, (tokenization, params) in enumerate(tok_params_sets):
-        tokenizer: miditok.MIDITokenizer = getattr(miditok, tokenization)(
-            tokenizer_config=miditok.TokenizerConfig(**params)
-        )
-
-        # Process the MIDI
-        # preprocess_midi is also performed when tokenizing, but we need to call it here
-        # for following adaptations
-        midi_to_compare = prepare_midi_for_tests(midi, tokenizer=tokenizer)
-        # Store preprocessed track
-        if len(tracks_with_errors) == 0:
-            tracks_with_errors += midi_to_compare.tracks
-            for ti, track in enumerate(midi_to_compare.tracks):
-                track.name = f"original {ti} quantized"
-
-        # printing the tokenizer shouldn't fail
-        _ = str(tokenizer)
-
-        # MIDI -> Tokens -> MIDI
-        decoded_midi, has_errors = tokenize_and_check_equals(
-            midi_to_compare, tokenizer, tok_i, midi_path.stem
-        )
-
-        # Add track to error list
-        if has_errors:
-            at_least_one_error = True
-            for track in decoded_midi.tracks:
-                track.name = f"{tok_i} encoded with {tokenization}"
-            tracks_with_errors += decoded_midi.tracks
-
-    # > 1 as the first one is the preprocessed
-    if len(tracks_with_errors) > len(midi.tracks) and saving_erroneous_midis:
-        TEST_LOG_DIR.mkdir(exist_ok=True, parents=True)
-        midi.tempo_changes = midi_to_compare.tempos
-        midi.time_signature_changes = midi_to_compare.time_signatures
-        midi.tracks += tracks_with_errors
-        midi.dump(TEST_LOG_DIR / midi_path.name)
-
-    assert not at_least_one_error
